refactor(core): strict C23/C++23 modernization & build hardening
- Enforce C23/C++23 standards in CMake & Gradle

- Enable 'Warnings as Errors' (-Werror -Wall -Wextra)

- Replace legacy TRUE/FALSE macros with standard bool

- Replace NULL with nullptr

- Fix implicit type conversions (size_t/int/char)

- Suppress unused legacy functions via [[maybe_unused]]

- Update documentation with new build standards

+++ app/src/main/jni/dlx.c
@@ -11,13 +11,14 @@
  */
 
 #include "dlx.h"
+
 #include <string.h>
 
 /*
  * Cover column c: Remove c from header list and remove all rows
  * that have a 1 in column c from their respective columns.
  */
-static void cover(dlx_node *c) {
+static void cover(dlx_node* c) {
     dlx_node *i, *j;
 
     /* Remove c from the horizontal header list */
@@ -40,7 +41,7 @@ static void cover(dlx_node *c) {
  * Uncover column c: Restore c and all its rows.
  * Must be called in exact reverse order of cover().
  */
-static void uncover(dlx_node *c) {
+static void uncover(dlx_node* c) {
     dlx_node *i, *j;
 
     /* For each row i in column c (in reverse order) */
@@ -63,12 +64,12 @@ static void uncover(dlx_node *c) {
  * Choose the column with minimum size (S heuristic).
  * This reduces branching factor and speeds up the search.
  */
-static dlx_node *choose_column(dlx_ctx *ctx) {
+static dlx_node* choose_column(dlx_ctx* ctx) {
     dlx_node *c, *best;
     int min_size;
 
     best = ctx->header->R;
-    if (best == ctx->header) return NULL;  /* Empty matrix */
+    if (best == ctx->header) return NULL; /* Empty matrix */
 
     min_size = best->size;
     for (c = best->R; c != ctx->header; c = c->R) {
@@ -85,7 +86,7 @@ static dlx_node *choose_column(dlx_ctx *ctx) {
  * Recursive search function.
  * Returns 1 when a solution is found, 0 otherwise.
  */
-static int search(dlx_ctx *ctx, int k) {
+static int search(dlx_ctx* ctx, int k) {
     dlx_node *c, *r, *j;
 
     /* If the matrix is empty, we found a solution */
@@ -98,7 +99,7 @@ static int search(dlx_ctx *ctx, int k) {
     /* Choose a column with minimum size */
     c = choose_column(ctx);
     if (!c || c->size == 0) {
-        return 0;  /* Dead end: column with no rows */
+        return 0; /* Dead end: column with no rows */
     }
 
     /* Cover the chosen column */
@@ -109,7 +110,7 @@ static int search(dlx_ctx *ctx, int k) {
         /* Add this row to the solution */
         if (k >= ctx->sol_cap) {
             int new_cap = ctx->sol_cap * 2;
-            int *new_sol = (int *)realloc(ctx->solution, new_cap * sizeof(int));
+            int* new_sol = (int*)realloc(ctx->solution, (size_t)new_cap * sizeof(int));
             if (!new_sol) return 0;
             ctx->solution = new_sol;
             ctx->sol_cap = new_cap;
@@ -141,15 +142,15 @@ static int search(dlx_ctx *ctx, int k) {
 /*
  * Create a new DLX context from a constraint matrix.
  */
-dlx_ctx *dlx_new(int n_rows, int n_cols, int **matrix) {
-    dlx_ctx *ctx;
+dlx_ctx* dlx_new(int n_rows, int n_cols, int** matrix) {
+    dlx_ctx* ctx;
     int i, j, node_count, node_idx;
-    int *row_first;  /* First node in each row */
-    int *row_last;   /* Last node in each row */
+    int* row_first; /* First node in each row */
+    int* row_last;  /* Last node in each row */
 
     if (n_rows <= 0 || n_cols <= 0 || !matrix) return NULL;
 
-    ctx = (dlx_ctx *)calloc(1, sizeof(dlx_ctx));
+    ctx = (dlx_ctx*)calloc(1, sizeof(dlx_ctx));
     if (!ctx) return NULL;
 
     ctx->n_cols = n_cols;
@@ -163,9 +164,9 @@ dlx_ctx *dlx_new(int n_rows, int n_cols, int **matrix) {
     }
 
     /* Allocate header + column headers */
-    ctx->header = (dlx_node *)calloc(1, sizeof(dlx_node));
-    ctx->columns = (dlx_node *)calloc(n_cols, sizeof(dlx_node));
-    ctx->nodes = (dlx_node *)calloc(node_count, sizeof(dlx_node));
+    ctx->header = (dlx_node*)calloc(1, sizeof(dlx_node));
+    ctx->columns = (dlx_node*)calloc((size_t)n_cols, sizeof(dlx_node));
+    ctx->nodes = (dlx_node*)calloc((size_t)node_count, sizeof(dlx_node));
 
     if (!ctx->header || !ctx->columns || !ctx->nodes) {
         dlx_destroy(ctx);
@@ -182,7 +183,7 @@ dlx_ctx *dlx_new(int n_rows, int n_cols, int **matrix) {
 
     /* Initialize column headers and link them horizontally */
     for (j = 0; j < n_cols; j++) {
-        dlx_node *col = &ctx->columns[j];
+        dlx_node* col = &ctx->columns[j];
         col->U = col->D = col;
         col->C = col;
         col->row = -1;
@@ -197,8 +198,8 @@ dlx_ctx *dlx_new(int n_rows, int n_cols, int **matrix) {
     }
 
     /* Track first/last node in each row for horizontal linking */
-    row_first = (int *)calloc(n_rows, sizeof(int));
-    row_last = (int *)calloc(n_rows, sizeof(int));
+    row_first = (int*)calloc((size_t)n_rows, sizeof(int));
+    row_last = (int*)calloc((size_t)n_rows, sizeof(int));
     if (!row_first || !row_last) {
         free(row_first);
         free(row_last);
@@ -213,10 +214,10 @@ dlx_ctx *dlx_new(int n_rows, int n_cols, int **matrix) {
     /* Create nodes and link vertically */
     node_idx = 0;
     for (j = 0; j < n_cols; j++) {
-        dlx_node *col = &ctx->columns[j];
+        dlx_node* col = &ctx->columns[j];
         for (i = 0; i < n_rows; i++) {
             if (matrix[i][j]) {
-                dlx_node *node = &ctx->nodes[node_idx];
+                dlx_node* node = &ctx->nodes[node_idx];
                 node->row = i;
                 node->C = col;
                 node->col_idx = j;
@@ -244,7 +245,7 @@ dlx_ctx *dlx_new(int n_rows, int n_cols, int **matrix) {
         if (row_first[i] >= 0) {
             dlx_node *first = NULL, *prev = NULL;
             for (j = 0; j < node_count; j++) {
-                dlx_node *node = &ctx->nodes[j];
+                dlx_node* node = &ctx->nodes[j];
                 if (node->row == i) {
                     if (!first) {
                         first = node;
@@ -268,7 +269,7 @@ dlx_ctx *dlx_new(int n_rows, int n_cols, int **matrix) {
 
     /* Allocate solution stack */
     ctx->sol_cap = n_rows > 16 ? n_rows : 16;
-    ctx->solution = (int *)malloc(ctx->sol_cap * sizeof(int));
+    ctx->solution = (int*)malloc((size_t)ctx->sol_cap * sizeof(int));
     if (!ctx->solution) {
         dlx_destroy(ctx);
         return NULL;
@@ -282,7 +283,7 @@ dlx_ctx *dlx_new(int n_rows, int n_cols, int **matrix) {
 /*
  * Free all resources.
  */
-void dlx_destroy(dlx_ctx *ctx) {
+void dlx_destroy(dlx_ctx* ctx) {
     if (!ctx) return;
     free(ctx->header);
     free(ctx->columns);
@@ -294,7 +295,7 @@ void dlx_destroy(dlx_ctx *ctx) {
 /*
  * Solve the exact cover problem.
  */
-int dlx_solve(dlx_ctx *ctx) {
+int dlx_solve(dlx_ctx* ctx) {
     if (!ctx) return 0;
     ctx->found = 0;
     ctx->sol_size = 0;
@@ -304,7 +305,7 @@ int dlx_solve(dlx_ctx *ctx) {
 /*
  * Get the solution.
  */
-int *dlx_get_solution(dlx_ctx *ctx, int *size) {
+int* dlx_get_solution(dlx_ctx* ctx, int* size) {
     if (!ctx || !ctx->found) {
         if (size) *size = 0;
         return NULL;
