+++ app/src/main/jni/maxflow_optimized.c
@@ -29,38 +29,38 @@
 #ifdef __ARM_NEON
 #include <arm_neon.h>
 #endif
-#include "maxflow.h"
-#include "puzzles.h"
 #include <assert.h>
 #include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
+
+#include "maxflow.h"
+#include "puzzles.h"
 
 /*
  * Fully Optimized maxflow.c with AVX2 Fast Path.
  */
 
 int maxflow_scratch_size(int nv) {
-    return (nv * 4) * sizeof(int);
+    return (int)(((size_t)nv * 4) * sizeof(int));
 }
 
-void maxflow_setup_backedges(int ne, const int *edges, int *backedges) {
+void maxflow_setup_backedges(int ne, const int* edges, int* backedges) {
     int i, n;
-    for (i = 0; i < ne; i++)
-        backedges[i] = i;
+    for (i = 0; i < ne; i++) backedges[i] = i;
     // Heapsort implementation (scalar, setup is not bottleneck)
     n = 0;
-#define LESS(i, j)                                                                                 \
-    ((edges[2 * (i) + 1] < edges[2 * (j) + 1]) ||                                                  \
+#define LESS(i, j)                                \
+    ((edges[2 * (i) + 1] < edges[2 * (j) + 1]) || \
      (edges[2 * (i) + 1] == edges[2 * (j) + 1] && edges[2 * (i)] < edges[2 * (j)]))
 #define PARENT(n) (((n) - 1) / 2)
 #define LCHILD(n) (2 * (n) + 1)
 #define RCHILD(n) (2 * (n) + 2)
-#define SWAP(i, j)                                                                                 \
-    do {                                                                                           \
-        int swaptmp = (i);                                                                         \
-        (i) = (j);                                                                                 \
-        (j) = swaptmp;                                                                             \
+#define SWAP(i, j)         \
+    do {                   \
+        int swaptmp = (i); \
+        (i) = (j);         \
+        (j) = swaptmp;     \
     } while (0)
     while (n < ne) {
         n++;
@@ -80,11 +80,9 @@ void maxflow_setup_backedges(int ne, const int *edges, int *backedges) {
         i = 0;
         while (1) {
             int lc = LCHILD(i), rc = RCHILD(i);
-            if (lc >= n)
-                break;
+            if (lc >= n) break;
             if (rc >= n) {
-                if (LESS(backedges[i], backedges[lc]))
-                    SWAP(backedges[i], backedges[lc]);
+                if (LESS(backedges[i], backedges[lc])) SWAP(backedges[i], backedges[lc]);
                 break;
             } else {
                 if (LESS(backedges[i], backedges[lc]) || LESS(backedges[i], backedges[rc])) {
@@ -102,12 +100,12 @@ void maxflow_setup_backedges(int ne, const int *edges, int *backedges) {
     }
 }
 
-int maxflow_with_scratch(void *scratch, int nv, int source, int sink, int ne, const int *edges,
-                         const int *backedges, const int *capacity, int *flow, int *cut) {
-    int *todo = (int *)scratch;
-    int *prev = todo + nv;
-    int *firstedge = todo + 2 * nv;
-    int *firstbackedge = todo + 3 * nv;
+int maxflow_with_scratch(void* scratch, int nv, int source, int sink, int ne, const int* edges,
+                         const int* backedges, const int* capacity, int* flow, int* cut) {
+    int* todo = (int*)scratch;
+    int* prev = todo + nv;
+    int* firstedge = todo + 2 * nv;
+    int* firstbackedge = todo + 3 * nv;
     int i, j, head, tail, from, to;
     int totalflow = 0;
 
@@ -115,38 +113,32 @@ int maxflow_with_scratch(void *scratch, int nv, int source, int sink, int ne, co
 
     j = 0;
     for (i = 0; i < ne; i++)
-        while (j <= edges[2 * i])
-            firstedge[j++] = i;
-    while (j < nv)
-        firstedge[j++] = ne;
+        while (j <= edges[2 * i]) firstedge[j++] = i;
+    while (j < nv) firstedge[j++] = ne;
 
     j = 0;
     for (i = 0; i < ne; i++)
-        while (j <= edges[2 * backedges[i] + 1])
-            firstbackedge[j++] = i;
-    while (j < nv)
-        firstbackedge[j++] = ne;
+        while (j <= edges[2 * backedges[i] + 1]) firstbackedge[j++] = i;
+    while (j < nv) firstbackedge[j++] = ne;
 
-    memset(flow, 0, ne * sizeof(int));
+    memset(flow, 0, (size_t)ne * sizeof(int));
 
     while (1) {
 #if defined(__AVX2__)
         __m256i neg_one = _mm256_set1_epi32(-1);
         for (i = 0; i < nv; i += 8) {
             if (i + 8 <= nv)
-                _mm256_storeu_si256((__m256i *)(prev + i), neg_one);
+                _mm256_storeu_si256((__m256i*)(prev + i), neg_one);
             else
-                for (int k = i; k < nv; k++)
-                    prev[k] = -1;
+                for (int k = i; k < nv; k++) prev[k] = -1;
         }
 #elif defined(__SSE2__) || defined(__x86_64__) || defined(__i386__)
         __m128i neg_one_sse = _mm_set1_epi32(-1);
         for (i = 0; i < nv; i += 4) {
             if (i + 4 <= nv)
-                _mm_storeu_si128((__m128i *)(prev + i), neg_one_sse);
+                _mm_storeu_si128((__m128i*)(prev + i), neg_one_sse);
             else
-                for (int k = i; k < nv; k++)
-                    prev[k] = -1;
+                for (int k = i; k < nv; k++) prev[k] = -1;
         }
 #elif defined(__ARM_NEON)
         int32x4_t neg_one_neon = vdupq_n_s32(-1);
@@ -154,12 +146,10 @@ int maxflow_with_scratch(void *scratch, int nv, int source, int sink, int ne, co
             if (i + 4 <= nv)
                 vst1q_s32(prev + i, neg_one_neon);
             else
-                for (int k = i; k < nv; k++)
-                    prev[k] = -1;
+                for (int k = i; k < nv; k++) prev[k] = -1;
         }
 #else
-        for (i = 0; i < nv; i++)
-            prev[i] = -1;
+        for (i = 0; i < nv; i++) prev[i] = -1;
 #endif
 
         uint32_t visited_mask = (1 << source);
@@ -180,8 +170,7 @@ int maxflow_with_scratch(void *scratch, int nv, int source, int sink, int ne, co
             }
             for (i = firstbackedge[from]; i < ne; i++) {
                 j = backedges[i];
-                if (edges[2 * j + 1] != from)
-                    break;
+                if (edges[2 * j + 1] != from) break;
                 to = edges[2 * j];
                 if (!(visited_mask & (1 << to))) {
                     if (flow[j] > 0) {
@@ -201,8 +190,7 @@ int maxflow_with_scratch(void *scratch, int nv, int source, int sink, int ne, co
                 int eidx = i / 2;
                 int spare =
                     (i & 1) ? flow[eidx] : (capacity[eidx] < 0 ? -1 : capacity[eidx] - flow[eidx]);
-                if (path_max < 0 || (spare >= 0 && spare < path_max))
-                    path_max = spare;
+                if (path_max < 0 || (spare >= 0 && spare < path_max)) path_max = spare;
                 to = edges[i];
             }
             to = sink;
